pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-ird-pricer"
version = "0.1.0"
description = "A lightweight, production-ready Flask web service and CLI for pricing plain-vanilla interest rate swaps"
authors = [{name = "Gemini IRD Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "Flask==3.1.0",
    "pandas==2.3.2",
    "numpy==2.0.2",
    "plotly==5.24.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "PyJWT==2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "hypothesis>=6.0.0",
]
metrics = [
    "prometheus-client==0.21.0",
]
security = [
    "bandit>=1.7.0",
]
production = [
    "gunicorn>=21.0.0",
    "prometheus-client==0.21.0",
]

[project.scripts]
gemini-ird-pricer = "gemini_ird_pricer.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/gemini-ird-pricer"
Repository = "https://github.com/your-org/gemini-ird-pricer"
Issues = "https://github.com/your-org/gemini-ird-pricer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_return_any = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=90"
python_files = ["tests/*.py"]
testpaths = ["tests"]
